refactor common fixtures
