+++ tests/test_metrics.py
@@ -1,12 +1,8 @@
 import base64
 import json
-from urllib.parse import urlparse
 
-from aiohttp import web
 from google.protobuf.json_format import MessageToDict
-import grpc.aio as grpc_aio
 from opentelemetry.proto.collector.metrics.v1.metrics_service_pb2 import ExportMetricsServiceRequest
-from opentelemetry.proto.collector.metrics.v1.metrics_service_pb2_grpc import MetricsServiceStub
 from opentelemetry.proto.common.v1.common_pb2 import AnyValue
 from opentelemetry.proto.common.v1.common_pb2 import KeyValue
 from opentelemetry.proto.metrics.v1.metrics_pb2 import Gauge
@@ -17,39 +13,10 @@ from opentelemetry.proto.metrics.v1.metrics_pb2 import ScopeMetrics
 from opentelemetry.proto.resource.v1.resource_pb2 import Resource
 import pytest
 
-from ddapm_test_agent.agent import DEFAULT_OTLP_GRPC_PORT
-from ddapm_test_agent.agent import DEFAULT_OTLP_HTTP_PORT
-from ddapm_test_agent.agent import make_otlp_grpc_server_async
-from ddapm_test_agent.client import TestOTLPClient
 from ddapm_test_agent.metrics import METRICS_ENDPOINT
-
-
-PROTOBUF_HEADERS = {"Content-Type": "application/x-protobuf"}
-JSON_HEADERS = {"Content-Type": "application/json"}
-
-
-async def _get_http_status_from_metadata(call):
-    """Extract HTTP status from GRPC trailing metadata."""
-    trailing_metadata = await call.trailing_metadata()
-    for key, value in trailing_metadata:
-        if key == "http-status":
-            return int(value)
-    return None
-
-
-@pytest.fixture
-def service_name():
-    return "ddservice"
-
-
-@pytest.fixture
-def environment():
-    return "ddenv"
-
-
-@pytest.fixture
-def version():
-    return "ddv1"
+from tests.conftest import JSON_HEADERS
+from tests.conftest import PROTOBUF_HEADERS
+from tests.conftest import _get_http_status_from_metadata
 
 
 @pytest.fixture
@@ -111,64 +78,6 @@ def otlp_metrics_string(otlp_metrics_protobuf):
 @pytest.fixture
 def otlp_metrics_json(otlp_metrics_protobuf):
     return json.dumps(MessageToDict(otlp_metrics_protobuf, preserving_proto_field_name=True))
-
-
-@pytest.fixture
-def otlp_http_url(testagent_url):
-    parsed_url = urlparse(testagent_url)
-    return f"{parsed_url.scheme}://{parsed_url.hostname}:{DEFAULT_OTLP_HTTP_PORT}"
-
-
-@pytest.fixture
-def otlp_test_client(otlp_http_url):
-    """OTLP client for retrieving stored metrics and requests."""
-    parsed_url = urlparse(otlp_http_url)
-    client = TestOTLPClient(parsed_url.hostname, parsed_url.port, parsed_url.scheme)
-    client.clear()
-    client.wait_to_start()
-    yield client
-    client.clear()
-
-
-@pytest.fixture
-async def otlp_grpc_client():
-    channel = grpc_aio.insecure_channel(f"127.0.0.1:{DEFAULT_OTLP_GRPC_PORT}")
-    stub = MetricsServiceStub(channel)
-
-    yield stub
-
-    await channel.close()
-
-
-@pytest.fixture
-async def grpc_server_with_failure_type(agent_app, available_port, aiohttp_server, request):
-    """GRPC server with configurable HTTP backend failure scenarios."""
-    grpc_port = int(available_port)
-    failure_type = getattr(request, "param", "connection_failure")
-
-    http_handlers = {
-        "http_400": lambda _: web.HTTPBadRequest(text="invalid"),
-        "http_500": lambda _: web.HTTPInternalServerError(text="boom"),
-    }
-
-    if failure_type == "connection_failure":
-        http_port = 99999  # Non-existent port
-    elif failure_type in http_handlers:
-        app = web.Application()
-        app.router.add_post(METRICS_ENDPOINT, http_handlers[failure_type])
-        http_server = await aiohttp_server(app)
-        http_port = http_server.port
-    else:
-        raise ValueError(f"Unknown failure_type: {failure_type}")
-
-    grpc_server = await make_otlp_grpc_server_async(agent_app.app["agent"], http_port=http_port, grpc_port=grpc_port)
-    channel = grpc_aio.insecure_channel(f"localhost:{grpc_port}")
-    stub = MetricsServiceStub(channel)
-
-    yield grpc_server, stub
-
-    await channel.close()
-    await grpc_server.stop(grace=0)
 
 
 async def test_metrics_endpoint_basic_http(testagent, otlp_http_url, otlp_metrics_string, loop):
@@ -384,9 +293,9 @@ async def test_metrics_endpoint_invalid_json(testagent, otlp_http_url, loop):
     assert resp.status == 400
 
 
-async def test_metrics_endpoint_basic_grpc(testagent, otlp_grpc_client, otlp_metrics_protobuf, loop):
+async def test_metrics_endpoint_basic_grpc(testagent, otlp_metrics_grpc_client, otlp_metrics_protobuf, loop):
     """GRPC metrics export with successful forwarding."""
-    call = otlp_grpc_client.Export(otlp_metrics_protobuf)
+    call = otlp_metrics_grpc_client.Export(otlp_metrics_protobuf)
     response = await call
 
     assert response is not None
@@ -401,7 +310,7 @@ async def test_metrics_endpoint_basic_grpc(testagent, otlp_grpc_client, otlp_met
 
 async def test_session_metrics_endpoint_grpc_forwarding(
     testagent,
-    otlp_grpc_client,
+    otlp_metrics_grpc_client,
     otlp_test_client,
     otlp_metrics_protobuf,
     service_name,
@@ -411,7 +320,7 @@ async def test_session_metrics_endpoint_grpc_forwarding(
     loop,
 ):
     """GRPC metrics forwarded to HTTP are retrievable via session endpoint."""
-    call = otlp_grpc_client.Export(otlp_metrics_protobuf)
+    call = otlp_metrics_grpc_client.Export(otlp_metrics_protobuf)
     response = await call
     assert response is not None
 
@@ -438,11 +347,11 @@ async def test_session_metrics_endpoint_grpc_forwarding(
     assert metrics_list[0]["name"] == metric_name
 
 
-@pytest.mark.parametrize("grpc_server_with_failure_type", ["http_400"], indirect=True)
-async def test_grpc_maps_http_400_to_metadata(grpc_server_with_failure_type):
+@pytest.mark.parametrize("service_grpc_server_with_failure_type", [("http_400", "metrics")], indirect=True)
+async def test_grpc_maps_http_400_to_metadata(service_grpc_server_with_failure_type):
     """GRPC forwarding preserves HTTP 400 status in metadata and partial_success."""
 
-    _, stub = grpc_server_with_failure_type
+    _, stub = service_grpc_server_with_failure_type
 
     call = stub.Export(ExportMetricsServiceRequest())
     response = await call
@@ -455,11 +364,11 @@ async def test_grpc_maps_http_400_to_metadata(grpc_server_with_failure_type):
     assert "HTTP 400" in response.partial_success.error_message
 
 
-@pytest.mark.parametrize("grpc_server_with_failure_type", ["http_500"], indirect=True)
-async def test_grpc_maps_http_500_to_metadata(grpc_server_with_failure_type):
+@pytest.mark.parametrize("service_grpc_server_with_failure_type", [("http_500", "metrics")], indirect=True)
+async def test_grpc_maps_http_500_to_metadata(service_grpc_server_with_failure_type):
     """GRPC forwarding preserves HTTP 500 status in metadata and partial_success."""
 
-    _, stub = grpc_server_with_failure_type
+    _, stub = service_grpc_server_with_failure_type
 
     call = stub.Export(ExportMetricsServiceRequest())
     response = await call
@@ -472,11 +381,11 @@ async def test_grpc_maps_http_500_to_metadata(grpc_server_with_failure_type):
     assert "HTTP 500" in response.partial_success.error_message
 
 
-@pytest.mark.parametrize("grpc_server_with_failure_type", ["connection_failure"], indirect=True)
-async def test_grpc_server_resilience_after_failure(grpc_server_with_failure_type, otlp_metrics_protobuf):
+@pytest.mark.parametrize("service_grpc_server_with_failure_type", [("connection_failure", "metrics")], indirect=True)
+async def test_grpc_server_resilience_after_failure(service_grpc_server_with_failure_type, otlp_metrics_protobuf):
     """GRPC server remains operational after processing failed requests."""
 
-    _, stub = grpc_server_with_failure_type
+    _, stub = service_grpc_server_with_failure_type
 
     call1 = stub.Export(otlp_metrics_protobuf)
     response1 = await call1
